// src/utils/types.ts
export interface ModelFile {
  rfilename: string;
  size?: number;
  url?: string;
  oid?: string;
  lfs?: {
    oid: string;
    size: number;
    pointerSize: number;
  };
  canFitInStorage?: boolean;
}